refactor: lnf에서 S3 기능 S3(global)로 적재
