+++ backend/src/main/java/com/passtival/backend/domain/lostfound/service/LnfService.java
@@ -9,7 +9,6 @@ import com.passtival.backend.domain.lostfound.model.response.FoundItemResponse;
 import com.passtival.backend.domain.lostfound.repository.LnfRepository;
 import com.passtival.backend.global.common.BaseResponseStatus;
 import com.passtival.backend.global.exception.BaseException;
-import com.passtival.backend.global.s3.S3Service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,11 +19,6 @@ import lombok.extern.slf4j.Slf4j;
 public class LnfService {
 
 	private final LnfRepository lnfRepository;
-	private final S3Service s3Service;
-
-	public String getUploadUrl(String fileName) {
-		return s3Service.generatePresignedUrl(fileName);
-	}
 
 	public FoundItemResponse getFoundItemById(Long id) {
 		FoundItem foundItem = lnfRepository.findById(id)
